Move stake-distribution properties from `TestChain` to a separate module
To make `TestChain` easier to navigate and increase cohesion
